Split rule implementation in false positive submodules
